Allow missing capabilities from source host
When loading a USB redirection stream during a qemu
migration, the source QEMU might be earlier and be missing
a bunch of capabilities that are now available in a more modern
version; allow this migration to work as long as the source
isn't claiming any capabilities that we don't have.

(We should be a bit more careful about this in future in qemu;
we could tie any new capabilities we ask for to machine types).

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>