usbredirtestclient: Add 64 bit id support
Some people are using usbredirtestclient as example code and cannot
get it to work with 64 bit ids, the cause of this is sending of
packets directly after calling usbredirparser_init without waiting
for the hello callback, which causes the packets to get send with
32 bit id headers, where as the receiver expects them with 64 bit ids
as it switches to 64 bit ids immediately after receiving the first
hello packet send by usbredirparser_init.

This commit adds 64 bit id support and moves the calling of
other packet sending functions after usbredirparser_init to the
hello callback to properly support 64 bit ids.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>