Avoid format truncation warnings on newer gcc
Newer versions  of gcc (e.g. 7.1.1 in fedora 26) print a warning about
format truncation even when using snprintf:

  CC       usbredirparser.lo
../../usbredirparser/usbredirparser.c: In function ‘usbredirparser_do_read’:
../../usbredirparser/usbredirparser.c:270:33: error: ‘%s’ directive output may be truncated writing up to 287 bytes into a region of size 64 [-Werror=format-truncation=]
     snprintf(buf, sizeof(buf), "%s", hello->version);
                                 ^~
../../usbredirparser/usbredirparser.c:270:5: note: ‘snprintf’ output between 1 and 288 bytes into a destination of size 64
     snprintf(buf, sizeof(buf), "%s", hello->version);
     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Because type_header is a 288-byte array that is cast to a 'struct
usb_redir_hello_header' and passed to the function, gcc apparently believes
that hello->version might be up to 288 bytes and warns about format truncation.
To avoid this warning, simply use strncpy (and ensure that the last byte is
NULL).