Don't print warning when resetting a missing device
When usbredirhost clears the current device, it resets the device and
attempts to re-attach the kernel driver for the device.  The current
device is cleared when the API user explicitly sets a new device with
usbredirhost_set_device().  However, some API users call
usbredirhost_set_device(..., NULL) to clear the current device when the
redirected device has been unplugged.  In this scenario, resetting the
device will fail and print a warning to the terminal. Since this is an
expected scenario, we should simply handle it rather than printing an
warning.

Resolves: rhbz#1442963
Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>